Refactor unit tests to Spock framework
